+++ src/OpenCVGenPkgConfig.info.cmake
@@ -17,7 +17,7 @@ set(OPENCV_INCLUDE_INSTALL_PATH "include/opencv-3.3.1-dev")
 
 set(OPENCV_3P_LIB_INSTALL_PATH "share/OpenCV-3.3.1-dev/3rdparty/lib")
 
-set(_modules "opencv_stitching;opencv_superres;opencv_videostab;opencv_aruco;opencv_bgsegm;opencv_bioinspired;opencv_ccalib;opencv_cvv;opencv_dpm;opencv_face;opencv_photo;opencv_fuzzy;opencv_hdf;opencv_img_hash;opencv_line_descriptor;opencv_optflow;opencv_reg;opencv_rgbd;opencv_saliency;opencv_stereo;opencv_structured_light;opencv_viz;opencv_phase_unwrapping;opencv_surface_matching;opencv_tracking;opencv_datasets;opencv_text;opencv_dnn;opencv_plot;opencv_xfeatures2d;opencv_shape;opencv_video;opencv_ml;opencv_ximgproc;opencv_calib3d;opencv_features2d;opencv_highgui;opencv_videoio;opencv_flann;opencv_xobjdetect;opencv_imgcodecs;opencv_objdetect;opencv_xphoto;opencv_imgproc;opencv_core")
+set(_modules "opencv_stitching;opencv_superres;opencv_videostab;opencv_aruco;opencv_bgsegm;opencv_bioinspired;opencv_ccalib;opencv_cvv;opencv_dpm;opencv_face;opencv_photo;opencv_freetype;opencv_fuzzy;opencv_hdf;opencv_img_hash;opencv_line_descriptor;opencv_optflow;opencv_reg;opencv_rgbd;opencv_saliency;opencv_stereo;opencv_structured_light;opencv_viz;opencv_phase_unwrapping;opencv_surface_matching;opencv_tracking;opencv_datasets;opencv_text;opencv_dnn;opencv_plot;opencv_xfeatures2d;opencv_shape;opencv_video;opencv_ml;opencv_ximgproc;opencv_calib3d;opencv_features2d;opencv_highgui;opencv_videoio;opencv_flann;opencv_xobjdetect;opencv_imgcodecs;opencv_objdetect;opencv_xphoto;opencv_imgproc;opencv_core")
 
 set(_extra "dl;m;pthread;rt")
 
@@ -44,6 +44,8 @@ set(TARGET_LOCATION_opencv_dpm "/home/pi/ros_catkin_ws/src/lib/libopencv_dpm3.so
 set(TARGET_LOCATION_opencv_face "/home/pi/ros_catkin_ws/src/lib/libopencv_face3.so.3.3.1")
 
 set(TARGET_LOCATION_opencv_photo "/home/pi/ros_catkin_ws/src/lib/libopencv_photo3.so.3.3.1")
+
+set(TARGET_LOCATION_opencv_freetype "/home/pi/ros_catkin_ws/src/lib/libopencv_freetype3.so.3.3.1")
 
 set(TARGET_LOCATION_opencv_fuzzy "/home/pi/ros_catkin_ws/src/lib/libopencv_fuzzy3.so.3.3.1")