+++ src/modules/freetype/CMakeFiles/opencv_freetype.dir/cmake_clean.cmake
@@ -1,0 +1,12 @@
+file(REMOVE_RECURSE
+  "CMakeFiles/opencv_freetype.dir/src/freetype.cpp.o"
+  "../../lib/libopencv_freetype3.pdb"
+  "../../lib/libopencv_freetype3.so"
+  "../../lib/libopencv_freetype3.so.3.3.1"
+  "../../lib/libopencv_freetype3.so.3.3"
+)
+
+# Per-language clean rules from dependency scanning.
+foreach(lang CXX)
+  include(CMakeFiles/opencv_freetype.dir/cmake_clean_${lang}.cmake OPTIONAL)
+endforeach()