+++ src/modules/freetype/CMakeFiles/opencv_freetype_pch_dephelp.dir/DependInfo.cmake
@@ -1,0 +1,40 @@
+# The set of languages for which implicit dependencies are needed:
+set(CMAKE_DEPENDS_LANGUAGES
+  "CXX"
+  )
+# The set of files for implicit dependencies of each language:
+set(CMAKE_DEPENDS_CHECK_CXX
+  "/home/pi/ros_catkin_ws/src/modules/freetype/opencv_freetype_pch_dephelp.cxx" "/home/pi/ros_catkin_ws/src/modules/freetype/CMakeFiles/opencv_freetype_pch_dephelp.dir/opencv_freetype_pch_dephelp.cxx.o"
+  )
+set(CMAKE_CXX_COMPILER_ID "GNU")
+
+# Preprocessor definitions for this target.
+set(CMAKE_TARGET_DEFINITIONS_CXX
+  "_USE_MATH_DEFINES"
+  "__OPENCV_BUILD=1"
+  "__STDC_CONSTANT_MACROS"
+  "__STDC_FORMAT_MACROS"
+  "__STDC_LIMIT_MACROS"
+  )
+
+# The include file search paths:
+set(CMAKE_CXX_TARGET_INCLUDE_PATH
+  "."
+  "/usr/include/eigen3"
+  "opencv3/opencv_contrib/freetype/include"
+  "opencv3/opencv_contrib/freetype/src"
+  "modules/freetype"
+  "opencv3/modules/core/include"
+  "opencv3/modules/imgproc/include"
+  "/usr/include/freetype2"
+  "/usr/include/harfbuzz"
+  "/usr/include/glib-2.0"
+  "/usr/lib/arm-linux-gnueabihf/glib-2.0/include"
+  )
+
+# Targets to which this target links.
+set(CMAKE_TARGET_LINKED_INFO_FILES
+  )
+
+# Fortran module output directory.
+set(CMAKE_Fortran_TARGET_MODULE_DIR "")