+++ src/modules/freetype/CMakeFiles/opencv_freetype_pch_dephelp.dir/cmake_clean_target.cmake
@@ -1,0 +1,3 @@
+file(REMOVE_RECURSE
+  "../../lib/libopencv_freetype_pch_dephelp.a"
+)