+++ src/modules/videoio/CMakeFiles/opencv_videoio.dir/cmake_clean.cmake
@@ -3,6 +3,7 @@ file(REMOVE_RECURSE
   "CMakeFiles/opencv_videoio.dir/src/cap_images.cpp.o"
   "CMakeFiles/opencv_videoio.dir/src/cap_mjpeg_encoder.cpp.o"
   "CMakeFiles/opencv_videoio.dir/src/cap_mjpeg_decoder.cpp.o"
+  "CMakeFiles/opencv_videoio.dir/src/cap_gstreamer.cpp.o"
   "CMakeFiles/opencv_videoio.dir/src/cap_v4l.cpp.o"
   "CMakeFiles/opencv_videoio.dir/src/cap_ffmpeg.cpp.o"
   "../../lib/libopencv_videoio3.pdb"