+++ src/ros/rosclean/CHANGELOG.rst
@@ -2,6 +2,14 @@
 Changelog for package rosclean
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+1.14.6 (2019-03-18)
+-------------------
+
+1.14.5 (2019-03-04)
+-------------------
+* enable rosclean on Windows (`#198 <https://github.com/ros/ros/issues/198>`_)
+* support BusyBox du (`#185 <https://github.com/ros/ros/issues/185>`_)
+
 1.14.4 (2018-05-01)
 -------------------