+++ src/rospack/CHANGELOG.rst
@@ -2,6 +2,10 @@
 Changelog for package rospack
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+2.4.5 (2019-06-07)
+------------------
+* add ros_environment as dependency (`#97 <https://github.com/ros/rospack/issues/97>`_)
+
 2.4.4 (2017-11-16)
 ------------------
 * workaround console_bridge macros (`#85 <https://github.com/ros/rospack/issues/85>`_)